engine + point-of-entry + wayworn-hush: the last two jobs get to finish, and both were failing on something other than the code -- CLANG-TIDY LINTS TESTS AND I HAD ONLY EVER RUN IT OVER src/, which is why thirty-eight findings in point-of-entry's own test files were news to CI and not to me; twenty of them were TEST_CASE bodies scored for cognitive complexity, which is measuring the wrong thing -- a case table SHOULD branch, and splitting one to satisfy a metric leaves a flood of tiny names that hide what is being tested -- and selva-oscura and prison-escape both already carry a tests/.clang-tidy saying exactly that, so point-of-entry simply never got the file every other game has; the other eighteen were real and are fixed: seven widening casts that multiplied as int before being widened (the rule tileIndex already follows), nine variables a test builds and only reads, a float loop counter, and an ifstream; two of my own fixes overreached onto things that ARE written and the compiler said so immediately, which is the argument for building after each pass; the ENGINE's PixelRenderTarget rounded by casting (x + 0.5) and widened a buffer size after multiplying, both now lround and a widened multiply; and CPPCHECK was failing on ONE finding in wayworn-hush that is not a fault at all -- a bare  inside a std::terminate handler is the idiom for getting a look at the exception that is still active, with the catch(...) beside it for when there is none, so it is suppressed AT THE SITE with the reason rather than in a list far away; every other cppcheck finding is style-severity and does not fail the job

+++ games/point-of-entry/tests/nav_test.cpp
@@ -18,7 +18,7 @@ EntityManager walledRoom()
     em.tile_map.height = 10;
     em.tile_map.tiles.assign(100, TileMap::Tile{0, true});
     const auto at = [&](int c, int r) -> TileMap::Tile&
-    { return em.tile_map.tiles[static_cast<std::size_t>(r * 10 + c)]; };
+    { return em.tile_map.tiles[static_cast<std::size_t>(r) * 10 + static_cast<std::size_t>(c)]; };
     for (int i = 0; i < 10; ++i)
     {
         at(i, 0).walkable = false;
@@ -140,7 +140,10 @@ TEST_CASE("an arch only takes a wall with body behind it", "[nav]")
     em.tile_map.height = 8;
     em.tile_map.tiles.assign(64, TileMap::Tile{0, true});
     const auto solid = [&](int c, int r)
-    { em.tile_map.tiles[static_cast<std::size_t>(r * 8 + c)].walkable = false; };
+    {
+        em.tile_map.tiles[static_cast<std::size_t>(r) * 8 + static_cast<std::size_t>(c)].walkable =
+            false;
+    };
 
     // A boundary at the top: two rows of rock, the way a carved room backs onto the mass.
     for (int c = 0; c < 8; ++c)
@@ -187,7 +190,10 @@ TEST_CASE("a spot for an arch is found at a wall, not merely on floor", "[nav]")
     em.tile_map.height = 10;
     em.tile_map.tiles.assign(100, TileMap::Tile{0, true});
     const auto solid = [&](int c, int r)
-    { em.tile_map.tiles[static_cast<std::size_t>(r * 10 + c)].walkable = false; };
+    {
+        em.tile_map.tiles[static_cast<std::size_t>(r) * 10 + static_cast<std::size_t>(c)].walkable =
+            false;
+    };
     // Rock across the top two rows; the rest of the map is open.
     for (int c = 0; c < 10; ++c)
     {
@@ -271,7 +277,7 @@ TEST_CASE("a slab's footprint is the whole connected block", "[nav]")
     em.tile_map.height = 12;
     em.tile_map.tiles.assign(144, TileMap::Tile{0, true});
     const auto at = [&](int c, int r) -> TileMap::Tile&
-    { return em.tile_map.tiles[static_cast<std::size_t>(r * 12 + c)]; };
+    { return em.tile_map.tiles[static_cast<std::size_t>(r) * 12 + static_cast<std::size_t>(c)]; };
     for (int i = 0; i < 12; ++i)
     {
         at(i, 0).walkable = false;
